got all the pages in and all the content
