+++ src/locale/translations/index.js
@@ -1,1 +1,1 @@
-export default JSON.parse("{\"en\":{\"Home\":{\"NoExposureDetected\":\"No exposure detected\",\"NoExposureDetectedDetailed\":\"You have not been near anyone who reported a COVID diagnosis through this app.\",\"ExposureDetected\":\"You’ve been exposed\",\"ExposureDetectedDetailed\":\"You’ve been near someone who reported a COVID-19 diagnosis through the app.\\n\\nThis does not mean you’ve been infected. But you should probably take some action.\",\"SeeGuidance\":\"What now? See guidance\",\"GuidanceUrl\":\"https://www.canada.ca/en/public-health/services/publications/diseases-conditions/covid-19-how-to-isolate-at-home.html\",\"SignalDataShared\":\"Thank you for helping to slow the spread\",\"SignalDataSharedDetailed\":{\"One\":\"For the next {number} day, you’ll get a daily message asking for your  information.\\n\\nTake care for yourself and your family during this time. \",\"Other\":\"For the next {number} days, you’ll get a daily message asking for your  information.\\n\\nTake care for yourself and your family during this time.\"},\"SignalDataSharedCTA\":\"Track your symptoms\",\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DailyShare\":\"Share your new information\",\"DailyShareDetailed\":\"Sharing your information helps others to know if they've been exposed to COVID-19. Then they can isolate and slow the spread to others. No details about who you are will be shared.\",\"ShareRandomIDsCTA\":\"Notify people you've been near\",\"BluetoothDisabled\":\"Bluetooth is off\",\"EnableBluetoothCTA\":\"You need to turn on Bluetooth in your phone’s settings so StopCOVID can work.\",\"TurnOnBluetooth\":\"Turn on Bluetooth\",\"ExposureNotificationsDisabled\":\"Exposure notifications are off\",\"ExposureNotificationsDisabledDetailed\":\"Exposure notifications are needed for StopCOVID to work.\",\"EnableExposureNotificationsCTA\":\"Turn on exposure notifications\",\"NoConnectivity\":\"No internet connection\",\"NoConnectivityDetailed\":\"StopCOVID needs internet access to continue checking for potential exposure.\",\"AppName\":\"StopCOVID\",\"ExternalLinkHint\":\"Opens in a new window\",\"LastCheckedMinutes\":{\"One\":\"Last checked {number} minute ago\",\"Other\":\"Last checked {number} minutes ago\"},\"LastCheckedHours\":{\"One\":\"Last checked {number} hour ago\",\"Other\":\"Last checked {number} hours ago\"},\"LastCheckedDays\":{\"One\":\"Last checked {number} day ago\",\"Other\":\"Last checked {number} days ago\"}},\"OverlayClosed\":{\"SystemStatus\":\"StopCOVID is \",\"SystemStatusOn\":\"ON\",\"SystemStatusOff\":\"OFF\",\"NotificationStatus\":\"Push notifications are \",\"NotificationStatusOff\":\"OFF\",\"TapPrompt\":\"Tap for more information\"},\"OverlayOpen\":{\"BluetoothCardStatus\":\"Bluetooth is \",\"BluetoothCardStatusOff\":\"OFF\",\"BluetoothCardBody\":\"You need to turn on Bluetooth in your phone’s settings so StopCOVID can work.\",\"BluetoothCardAction\":\"Turn on Bluetooth\",\"EnterCodeCardTitle\":\"Diagnosed with COVID-19?\",\"EnterCodeCardBody\":\"You were given an 8-digit number when you were diagnosed. This number tells StopCOVID to notify people you were near.\",\"EnterCodeCardAction\":\"Enter code\",\"NotificationCardStatus\":\"Push notifications are \",\"NotificationCardStatusOff\":\"OFF\",\"NotificationCardBody\":\"You will not receive push notifications if you've been near anyone who has reported a COVID-19 diagnosis. Finding out potential exposures right away is essential to slowing the spread of COVID-19.\",\"NotificationCardAction\":\"Turn on push notifications\",\"ExposureNotificationCardStatus\":\"Exposure notifications are \",\"ExposureNotificationCardStatusOff\":\"OFF\",\"ExposureNotificationCardBody\":\"Exposure notifications are needed for StopCOVID to work.\",\"ExposureNotificationCardAction\":\"Turn on exposure notifications\"},\"Info\":{\"CheckSymptoms\":\"Check symptoms\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/en\",\"TellAFriend\":\"Share app\",\"LearnMore\":\"How StopCOVID works\",\"ChangeLanguage\":\"Change language\",\"Privacy\":\"Read privacy policy\"},\"LanguageSelect\":{\"Title\":\"Language\",\"En\":\"English (Canada)\",\"Fr\":\"Français (Canada)\",\"PtBR\":\"Portuguese (Brazil)\",\"EnShort\":\"English\",\"FrShort\":\"Français\",\"PtBRShort\":\"Portuguese (Brazil)\",\"Close\":\"Back\"},\"Tutorial\":{\"step-1Title\":\"How StopCOVID works\",\"step-1\":\"Your phone uses Bluetooth to collect and share random IDs with phones near you with StopCOVID installed. These IDs are stored securely on each phone.\",\"step-2Title\":\"If you have potentially been exposed to COVID-19\",\"step-2\":\"You will receive a notification if you have been near someone in the past 14 days who has tested positive for COVID-19.\",\"step-3Title\":\"If you test positive for COVID-19\",\"step-3\":\"Anonymously notify others who have been near you in the past 14 days that they have possibly been exposed.\",\"ActionBack\":\"Back\",\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"Close\":\"Close\"},\"Sharing\":{\"Title\":\"Tell your friends about StopCOVID\",\"SubTitle\":\"Every person using StopCOVID helps to slow the spread of COVID-19. Sharing this app will not share any of your personal information.\",\"Platform-messages\":\"Share to Messages\",\"Platform-instagram\":\"Share to Instagram\",\"More\":\"More apps\",\"Message\":\"Join me in helping to slow the spread of COVID-19. Download the StopCOVID app: https://covidshield.app\",\"InstagramImageUrl\":\"https://user-images.githubusercontent.com/5274722/84658989-b9ba6b00-aee4-11ea-84d4-d840527467b4.png\",\"Close\":\"Close\"},\"Onboarding\":{\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"ActionBack\":\"Back\",\"Start\":{\"Title\":\"Join the effort to slow the spread\",\"Body1\":\"COVID Alert notifies you if someone you’ve been near tells the app they tested positive.\",\"Body2\":\"That way, you can isolate right away and stop the spread.\",\"TutorialAction\":\"Learn how StopCOVID works\"},\"Anonymous\":{\"Title\":\"Totally Anonymous\",\"Body1\":\"COVID Alert does not use GPS or location services.\",\"Body2\":\"It has no way of knowing:\",\"Bullet1\":\"your location\",\"Bullet2\":\"your name or address\",\"Bullet3\":\"your phone's contacts\",\"Bullet4\":\"your health information\"},\"HowItWorks\":{\"Title\":\"How it works\",\"Body1\":\"The app uses Bluetooth to exchange random codes with nearby phones.\",\"Body2\":\"Every day, it checks a list of random codes from people who tell the app they tested positive.\",\"Body3\":\"If you’ve been near one of those codes in the past 14 days, you’ll get a notification.\",\"HowItWorksCTA\":\"Learn more about how it works\"},\"Permissions\":{\"Title\":\"The app will now ask your permission\",\"Body1\":\"Allow the app to start logging random codes or “random IDs” when you’re near other phones. The app will access the date, duration and signal strength related to the random codes, but they never leave your phone.\",\"Body2\":\"You’ll also need to enable push notifications.\",\"PrivacyButtonCTA\":\"Learn more about privacy\"}},\"Privacy\":{\"Title\":\"Privacy policy\",\"Close\":\"Close\"},\"ThankYou\":{\"Title\":\"Thank you for helping\",\"Body\":\"Exposure notifications are on. You will receive a message if StopCOVID detects that you've been near someone who has reported a COVID-19 diagnosis.\",\"Dismiss\":\"Dismiss\"},\"DataUpload\":{\"Cancel\":\"Cancel\",\"FormIntro\":\"Please enter your 8 digit StopCOVID code.\",\"Action\":\"Submit code\",\"InputLabel\":\"StopCOVID code\",\"InfoSmall\":\"Your random IDs will not be shared unless you give permission in the next step.\",\"ConsentTitle\":\"Share your random IDs\",\"ConsentBody\":\"You have been granted access by a health care professional to share the random IDs stored on your phone.\",\"ConsentBody2\":\"If you share your random IDs, others with StopCOVID who have been near you in the past 14 days will receive a notification that they have possibly been exposed. This notification will not include any information about you.\",\"ConsentBody3\":\"You will be notified once per day when new random IDs are available to be shared. You must grant permission each time for the random IDs to be shared. Sharing is voluntary.\",\"PrivacyPolicyLink\":\"View privacy policy\",\"ConsentAction\":\"I agree\",\"ShareToast\":\"Random IDs shared successfully\",\"ErrorTitle\":\"Code not recognized\",\"ErrorBody\":\"Your StopCOVID code could not be recognized. Please try again.\",\"ErrorAction\":\"OK\"},\"Notification\":{\"ExposedMessageTitle\":\"You’ve been exposed\",\"ExposedMessageBody\":\"You’ve been near someone who reported a COVID-19 diagnosis through the app.\\n\\nThis does not mean you’ve been infected. But you should probably take some action.\",\"OffMessageTitle\":\"StopCOVID is off\",\"OffMessageBody\":\"Turn on StopCOVID to get notified if you've been near someone who has reported a COVID-19 diagnosis.\",\"DailyUploadNotificationTitle\":\"Share your new informatin about people you were near\",\"DailyUploadNotificationBody\":\"Sharing your information helps people know if they've been exposed to COVID-19. Then they can isolate and slow the spread to others. No details about who you are will be shared.\"},\"Partners\":{\"Label\":\"Developed in partnership with\"},\"BottomSheet\":{\"Collapse\":\"Close\",\"OnStatus\":\"StopCOVID is on, Tap for more information\",\"OffStatus\":\"StopCOVID is off, Tap for more information\"},\"RegionPicker\":{\"Title\":\"Where do you live?\",\"Body\":\"Choose the province or territory whose public health guidance you need to follow.\\n\\nThis information is only stored on your phone and will never be shared with anyone else.\",\"Skip\":\"Skip\",\"GetStarted\":\"Get started\",\"AB\":\"Alberta\",\"BC\":\"British Columbia\",\"MB\":\"Manitoba\",\"NB\":\"New Brunswick\",\"NL\":\"Newfoundland and Labrador\",\"NT\":\"Northwest Territories\",\"NS\":\"Nova Scotia\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Prince Edward Island\",\"QC\":\"Quebec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\"}},\"fr\":{\"Home\":{\"NoExposureDetected\":{\"NoRegion\":{\"Body\":\"Vous n’avez été près d’aucune personne ayant signalé un diagnostic de COVID-19 avec cette application.\\n\\nToutefois, dans certaines provinces et certains territoires, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\",\"DiagnosedBtnText1\":\"Vous avez la COVID-19?\",\"DiagnosedBtnText2\":\"Prochaines étapes\",\"Title\":\"Aucune exposition détectée\"},\"RegionCovered\":{\"Body\":\"Vous n’avez été près d’aucune personne ayant signalé un diagnostic de COVID-19 avec cette application.\",\"DiagnosedBtnText1\":\"Vous avez la COVID-19?\",\"DiagnosedBtnText2\":\"Prochaines étapes\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"Title\":\"Aucune exposition détectée\"},\"RegionNotCovered\":{\"Body\":\"Les personnes de votre province ou territoire ne peuvent pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nIl est tout de même utile d’activer Alerte COVID. Quand les personnes pourront signaler un diagnostic, vous serez ainsi avisé si jamais vous avez été près d’elles.\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"Title\":\"Pas encore de signalement dans votre région\"}},\"NoExposureDetectedDetailed\":\"You have not been near anyone who reported a COVID diagnosis through this app.\",\"ExposureDetected\":{\"Detailed1\":\"Vous avez été près d’une personne qui a signalé un diagnostic de COVID-19 avec cette application.\",\"Detailed2\":\"Cela ne veut pas dire que vous êtes infecté. Mais vous devriez probablement prendre quelques mesures.\",\"DiagnosedBtnText1\":\"Vous avez la COVID-19?\",\"DiagnosedBtnText2\":\"Prochaines étapes\",\"Title\":\"Vous avez été exposé\"},\"ExposureDetectedDetailed\":\"You’ve been near someone who reported a COVID-19 diagnosis through the app.\\n\\nThis does not mean you’ve been infected. But you should probably take some action.\",\"SeeGuidance\":\"Que faire maintenant? Lisez les recommandations.\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"SignalDataShared\":\"Thank you for helping to slow the spread\",\"SignalDataSharedDetailed\":{\"One\":\"For the next {number} day, you’ll get a daily message asking for your  information.\\n\\nTake care for yourself and your family during this time. \",\"Other\":\"For the next {number} days, you’ll get a daily message asking for your  information.\\n\\nTake care for yourself and your family during this time.\"},\"SignalDataSharedCTA\":\"Track your symptoms\",\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DailyShare\":\"Partagez vos nouveaux identifiants aléatoires\",\"DailyShareDetailed\":\"Partager vos données aide les autres à savoir s'ils ont possiblement été exposés à la COVID-19.\",\"ShareRandomIDsCTA\":\"Partagez vos identifiants aléatoires\",\"BluetoothDisabled\":\"Bluetooth désactivé\",\"EnableBluetoothCTA\":\"Vous devez activer Bluetooth dans les réglages de votre téléphone pour que Alerte COVID fonctionne.\",\"TurnOnBluetooth\":\"Activez Bluetooth\",\"ExposureNotificationsDisabled\":\"Les notifications d'exposition à la COVID-19 sont désactivées\",\"ExposureNotificationsDisabledDetailed\":\"Les notifications d'exposition à la COVID-19 sont nécessaires au fonctionnement de Alerte COVID.\",\"EnableExposureNotificationsCTA\":\"Activez les notifications d'exposition à la COVID-19\",\"NoConnectivity\":\"Alerte COVID est hors ligne\",\"NoConnectivityDetailed\":\"Alerte COVID vérifie s’il y a des expositions seulement quand votre téléphone est connecté à Internet.\",\"AppName\":\"Alerte COVID\",\"ExternalLinkHint\":\"Ouvre dans une nouvelle fenêtre\",\"LastCheckedMinutes\":{\"One\":\"Dernière vérification il y a {number} minute\",\"Other\":\"Dernière vérification il y a {number} minutes\"},\"LastCheckedHours\":{\"One\":\"Dernière vérification il y a {number} heure\",\"Other\":\"Dernière vérification il y a {number} heures\"},\"LastCheckedDays\":{\"One\":\"Dernière vérification il y a {number} jour\",\"Other\":\"Dernière vérification il y a {number} jours\"},\"CTA\":\"Prochaines étapes\",\"ChooseRegionCTA\":\"Sélectionnez une province ou un territoire\",\"DiagnosedView\":{\"Body1\":{\"One\":\"Tomorrow, you’ll get your last notification asking permission to upload any new random codes your phone sent out.\",\"Other\":\"Au cours des {number} prochains jours, vous recevrez une notification quotidienne vous demandant l’autorisation de téléverser tout code aléatoire émis par votre téléphone.\"},\"Body2\":\"Vous n’êtes pas tenu d’accepter, même si vous avez accepté le jour d’avant. Vous avez le choix.\",\"Body3\":\"Prenez soin de vous et de votre famille pendant ce temps.\",\"Title\":\"Merci d’aider à lutter contre la COVID\"},\"How\":\"Diagnosed with COVID-19?\",\"HowUrl\":\"https://www.canada.ca/\"},\"OverlayClosed\":{\"SystemStatus\":\"Alerte COVID est \",\"SystemStatusOn\":\"ACTIVÉ\",\"SystemStatusOff\":\"DÉSACTIVÉ\",\"NotificationStatus\":\"Notifications \",\"NotificationStatusOff\":\"DÉSACTIVÉES\",\"TapPrompt\":\"Appuyez pour voir le menu\"},\"OverlayOpen\":{\"BluetoothCardStatus\":\"Bluetooth \",\"BluetoothCardStatusOff\":\"DÉSACTIVÉ\",\"BluetoothCardBody\":\"Vous devez l’activer dans les réglages de votre téléphone pour que Alerte COVID fonctionne.\",\"BluetoothCardAction\":\"Activer Bluetooth\",\"EnterCodeCardTitle\":\"Code Alerte COVID\",\"EnterCodeCardBody\":\"Un professionel de la santé vous a-t-il demandé d'entrer un code Alerte COVID?\",\"EnterCodeCardAction\":\"Entrez le code\",\"NotificationCardStatus\":\"Activer les notifications \",\"NotificationCardStatusOff\":\"DÉSACTIVÉES\",\"NotificationCardBody\":\"Vous ne recevrez pas de notifications poussées en cas d’exposition potentielle à la COVID-19. Le fait d’être rapidement notifié d’une exposition potentielle est important pour ralentir la propagation de COVID-19.\",\"NotificationCardAction\":\"Activer les notifications\",\"ExposureNotificationCardStatus\":\"Les notifications d'exposition à la COVID-19 sont \",\"ExposureNotificationCardStatusOff\":\"DÉSACTIVÉES\",\"ExposureNotificationCardBody\":\"Les notifications d'exposition à la COVID-19 sont nécessaires au fonctionnement de Alerte COVID.\",\"ExposureNotificationCardAction\":\"Activez les notifications d'exposition\",\"EnterCodeCardBodyDiagnosed\":\"Merci de faire votre part pour freiner la propagation du virus.\",\"EnterCodeCardTitleDiagnosed\":\"Vous contribuez à freiner la COVID\"},\"Info\":{\"CheckSymptoms\":\"Vérifier vos symptômes\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/fr\",\"TellAFriend\":\"Partager l’application\",\"LearnMore\":\"Fonctionnement de Alerte COVID\",\"ChangeLanguage\":\"Modifier la langue\",\"Privacy\":\"Lire la politique de confidentialité\",\"ChangeRegion\":\"Modifier la province\"},\"LanguageSelect\":{\"Title\":\"Langue\",\"En\":\"English (Canada)\",\"Fr\":\"Français (Canada)\",\"PtBR\":\"Portugais (Brésil)\",\"EnShort\":\"English\",\"FrShort\":\"Français\",\"PtBRShort\":\"Portugais (Brésil)\",\"Close\":\"Retour\"},\"Tutorial\":{\"step-1Title\":\"Fonctionnement de Alerte COVID\",\"step-1\":\"Après son installation, Alerte COVID s’exécute en arrière-plan sans interrompre vos activités.\\n\\nAlors que vous vaquez à vos activités, votre téléphone émet un code aléatoire par Bluetooth toutes les 5 minutes. L’application change fréquemment les codes. Ainsi, ils ne peuvent pas servir à vous identifier.\",\"step-2Title\":\"\",\"step-2\":\"Chaque fois qu’une personne ayant installé Alerte COVID se trouve près de vous, son téléphone émet des codes aléatoires que votre téléphone collecte et stocke. De même, l’autre téléphone collecte et stocke les codes aléatoires émis par votre téléphone.\\n\\nLes téléphones émettent des codes aléatoires toutes les 5 minutes. Ces codes sont stockés sur chacun des téléphones de façon sécurisée.\",\"step-3Title\":\"\",\"step-3\":\"L’application estime à quelle distance se trouvent les autres téléphones ayant Alerte COVID. Pour ce faire, elle utilise la puissance de leur signal Bluetooth.\\n\\nSi vous êtes à moins d’une certaine distance pendant plus d’un certain temps, l’application enregistre une exposition.\",\"ActionBack\":\"Précédent\",\"ActionNext\":\"Suivant\",\"ActionEnd\":\"Terminé\",\"Close\":\"Fermer\",\"step-4\":\"Si une personne ayant l’application a un diagnostic de COVID-19, elle peut décider de téléverser tous les codes émis par son téléphone. Les codes vont vers un serveur central.\\n\\nLe serveur ne reçoit que les codes. Il ne reçoit aucune information sur la personne.\",\"step-4Title\":\"\",\"step-5\":\"Chaque jour, votre téléphone vérifie le serveur quand il a une connexion Internet. Il obtient une liste de codes aléatoires émis par les téléphones des personnes qui ont signalé un diagnostic avec l’application.\\n\\nL’application supprime les codes collectés après 14 jours.\",\"step-5Title\":\"\",\"step-6\":\"L’application compare la liste de codes avec les codes aléatoires reçus d’autres téléphones.\\n\\nSi elle trouve des codes identiques, elle vérifie la puissance du signal Bluetooth et la durée pendant laquelle vous avez été à proximité. Si les codes sont associés à une exposition, l’application vous dira que vous avez peut-être été exposé.\\n\\nVous pouvez alors vous isoler et lire les recommandations de santé publique.\",\"step-6Title\":\"\"},\"Sharing\":{\"Title\":\"Partager l’application\",\"SubTitle\":\"Chaque personne utilisant Alerte COVID aide à ralentir la transmission de la COVID-19. Partager cette application ne partage aucune donnée privée.\",\"Platform-messages\":\"Partager avec Messages\",\"Platform-instagram\":\"Partager avec Instagram\",\"More\":\"Plus d'applications\",\"Message\":\"Joignez-vous à moi pour aider à ralentir la transmission de la COVID-19. Téléchargez l'application Alerte COVID : https://covidshield.app\",\"InstagramImageUrl\":\"https://user-images.githubusercontent.com/5274722/84658989-b9ba6b00-aee4-11ea-84d4-d840527467b4.png\",\"Close\":\"Fermer\"},\"Onboarding\":{\"ActionNext\":\"Suivant\",\"ActionEnd\":\"Terminé\",\"ActionBack\":\"Précédent\",\"Start\":{\"Title\":\"Join the effort to slow the spread\",\"Body1\":\"COVID Alert notifies you if someone you’ve been near tells the app they tested positive.\",\"Body2\":\"That way, you can isolate right away and stop the spread.\",\"TutorialAction\":\"Learn how StopCOVID works\"},\"Anonymous\":{\"Title\":\"Totally Anonymous\",\"Body1\":\"COVID Alert does not use GPS or location services.\",\"Body2\":\"It has no way of knowing:\",\"Bullet1\":\"your location\",\"Bullet2\":\"your name or address\",\"Bullet3\":\"your phone's contacts\",\"Bullet4\":\"your health information\"},\"HowItWorks\":{\"Title\":\"How it works\",\"Body1\":\"The app uses Bluetooth to exchange random codes with nearby phones.\",\"Body2\":\"Every day, it checks a list of random codes from people who tell the app they tested positive.\",\"Body3\":\"If you’ve been near one of those codes in the past 14 days, you’ll get a notification.\",\"HowItWorksCTA\":\"Learn more about how it works\"},\"Permissions\":{\"Title\":\"The app will now ask your permission\",\"Body1\":\"Allow the app to start logging random codes or “random IDs” when you’re near other phones. The app will access the date, duration and signal strength related to the random codes, but they never leave your phone.\",\"Body2\":\"You’ll also need to enable push notifications.\",\"PrivacyButtonCTA\":\"Learn more about privacy\"}},\"Privacy\":{\"Title\":\"Politique de confidentialité\",\"Close\":\"Fermer\"},\"ThankYou\":{\"Title\":\"Merci d'aider\",\"Body\":\"Les notifications d’exposition sont activées. Si Alerte COVID détecte que vous avez possiblement été exposé(e) à la COVID-19, vous recevrez une notification.\",\"Dismiss\":\"OK\"},\"DataUpload\":{\"Cancel\":\"Annuler\",\"FormIntro\":\"Entrez votre numéro à 8 chiffres.\",\"Action\":\"Envoyer\",\"InputLabel\":\"code Alerte COVID\",\"InfoSmall\":\"Vos identifiants aléatoires ne seront pas partagés, à moins que vous donniez votre permission à l'étape suivante.\",\"ConsentTitle\":\"Permettre à Alerte COVID d’aviser les autres personnes\",\"ConsentBody\":\"You have been granted access by a health care professional to share the random IDs stored on your phone.\",\"ConsentBody2\":\"Personne ne recevra d’information sur vous ou sur le moment où vous avez été en proximité.\",\"ConsentBody3\":\"Il s’agit de votre choix. Vous pouvez aider à ralentir la propagation de COVID-19 si vous acceptez.\",\"PrivacyPolicyLink\":\"Comment Alerte COVID protège votre vie privée\",\"ConsentAction\":\"Accepter\",\"ShareToast\":\"Identifiants aléatoires partagés avec succès\",\"ErrorTitle\":\"Code non reconnu\",\"ErrorBody\":\"Votre code Alerte COVID n’a pas pu être reconnu. Veuillez essayer à nouveau.\",\"ErrorAction\":\"OK\",\"ConsentBody1\":\"Alerte COVID vous demandera l’autorisation de téléverser vos codes aléatoires des 14 derniers jours.\",\"ConsentSubtitle\":\"C’est votre choix\",\"HowItWorksLink\":\"Comment Alerte COVID fonctionne\",\"Step1\":{\"Body\":\"Vous devriez avoir reçu un numéro à 8 chiffres quand vous avez eu votre diagnostic. Vous devez entrer ce numéro si vous voulez que Alerte COVID avise les personnes qu’elles ont été exposées.\\n\\nEntrer le code est optionnel. Vous avez le choix. Mais vous pouvez aider à ralentir la propagation de COVID-19 si vous le faites.\",\"CTA\":\"Entrer le numéro\",\"NoCode\":\"Vous n’avez pas de numéro à 8 chiffres?\",\"Title\":\"Entrez votre numéro à 8 chiffres\"},\"NoCode\":{\"NoRegion\":{\"Title\":\"Vous n’avez pas choisi de province ou de territoire\",\"Body\":\"Dans certaines provinces et certains territoires, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nPour savoir si vous pouvez obtenir un numéro à 8 chiffres, vous devez sélectionner votre région.\",\"ChooseRegionCTA\":\"Sélectionner une province ou un territoire\"},\"RegionCovered\":{\"Title\":\"Vous n’avez pas de numéro à 8 chiffres?  \",\"Body\":\"Contactez la personne ou le bureau qui vous a donné vos résultats afin d’obtenir un numéro.\"},\"RegionNotCovered\":{\"Title\":\"Pas encore de signalement dans votre région\",\"Body\":\"Les personnes de votre province ou territoire ne peuvent pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nVous ne pourrez pas aviser les personnes dont vous avez été à proximité qu’elles ont été exposées. Mais vous pouvez tout de même prendre des mesures pour :\",\"Body2\":\"assurer la sécurité de votre communauté;\",\"Body3\":\"protéger les personnes et les animaux vivant avec vous;\",\"Body4\":\"prendre soin de vous.\"}}},\"Notification\":{\"ExposedMessageTitle\":\"Vous avez possiblement été exposé(e)\",\"ExposedMessageBody\":\"Selon vos identifiants aléatoires, au cours des 14 derniers jours, vous avez été à proximité d'une personne ayant reçu un résultat de test positif à la COVID-19.\",\"OffMessageTitle\":\"Alerte COVID est désactivé\",\"OffMessageBody\":\"Activez Alerte COVID pour recevoir une notification si vous avez potentiellement été exposé(e) à la COVID-19.\",\"DailyUploadNotificationTitle\":\"Partagez vos nouveaux identifiants aléatoires\",\"DailyUploadNotificationBody\":\"Partager vos données aide les autres à savoir s'ils ont possiblement été exposés à la COVID-19.\"},\"Partners\":{\"Label\":\"Conçu en partenariat avec\"},\"BottomSheet\":{\"Collapse\":\"Fermer\",\"OnStatus\":\"Alerte COVID est ACTIVÉ, Appuyez pour voir le menu\",\"OffStatus\":\"Alerte COVID est DÉSACTIVÉ, Appuyez pour voir le menu\"},\"RegionPicker\":{\"Title\":\"Où habitez-vous?\",\"Body\":\"Sélectionnez la province ou le territoire dont vous devez suivre les recommandations de santé publique.\\n\\nCette information est stockée sur votre téléphone uniquement et ne sera jamais transmise à d’autres personnes.\",\"Skip\":\"Ignorer\",\"GetStarted\":\"Commencer\",\"AB\":\"Alberta\",\"BC\":\"Colombie-Britannique\",\"MB\":\"Manitoba\",\"NB\":\"Nouveau-Brunswick\",\"NL\":\"Terre-Neuve-et-Labrador\",\"NT\":\"Territoires du Nord-Ouest\",\"NS\":\"Nouvelle-Écosse\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Île-du-Prince-Édouard\",\"QC\":\"Québec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\",\"Close\":\"Fermer\",\"Optional\":\"* Facultatif\",\"SettingsTitle\":\"Province ou territoire\"},\"OnboardingLocation\":{\"Body1\":\"votre position\",\"Body2\":\"votre identité\",\"Body3\":\"les endroits où vous êtes allé\",\"Body4\":\"les personnes avec qui vous avez été\",\"Body5\":\"Vous pouvez le supprimer à tout moment\",\"Intro\":\"Alerte COVID n’utilise pas le GPS ou les services de localisation. Il n’a aucun moyen de connaître :\",\"Title\":\"Aucun suivi de position\"},\"OnboardingPermissions\":{\"Body1\":\"Rien n’est fait sans votre autorisation.\",\"Body2\":\"Même quand vous donnez votre autorisation, Alerte COVID ne partage que les codes aléatoires des téléphones dont vous avez été à proximité. Il ne partage aucune information sur vous.\",\"HowItWorksCTA\":\"Lire sur le fonctionnement de Alerte COVID\",\"PrivacyButtonCTA\":\"Lire la politique de confidentialité\",\"Title\":\"Aucune action sans votre autorisation\"},\"OnboardingStart\":{\"Body1\":\"Il vous avise si au cours des 14 derniers jours, vous avez été à proximité d’une personne ayant signalé un diagnostic de COVID-19 avec l’application.\",\"Body2\":\"Il ne vous dira pas en temps réel si vous êtes à proximité d’une personne ayant eu un diagnostic.\",\"Body3\":\"If you test positive for COVID-19, you can anonymously notify others who have been near you, in the past 14 days.\",\"Body4\":\"Vous n’êtes pas tenu de l’utiliser\",\"Body5\":\"Vous pouvez le supprimer à tout moment\",\"Title\":\"Ce que fait Alerte COVID\",\"Title2\":\"Vous avez le choix\",\"TutorialAction\":\"Découvrez comment Alerte COVID fonctionne\"},\"RegionalGuidanceURL\":{\"AB\":\"\",\"BC\":\"\",\"CA\":\"https://www.canada.ca/fr/sante-publique/services/maladies/2019-nouveau-coronavirus/symptomes.html\",\"MB\":\"https://www.gov.mb.ca/covid19/about/index.fr.html\",\"NB\":\"https://www2.gnb.ca/content/gnb/fr/corporate/promo/covid-19.html\",\"NL\":\"\",\"NS\":\"https://novascotia.ca/coronavirus/when-to-seek-help/fr/\",\"NT\":\"https://www.gov.nt.ca/covid-19/fr/services/%C3%A0-propos-de-la-covid-19\",\"NU\":\"https://www.gov.nu.ca/fr/sante/information/covid-19-nouveau-coronavirus-0\",\"ON\":\"https://covid-19.ontario.ca/fr/centres-devaluation-de-la-covid-19\",\"PE\":\"https://www.princeedwardisland.ca/fr/information/sante-et-mieux-etre/covid-19-foire-aux-questions\",\"QC\":\"https://www.quebec.ca/sante/problemes-de-sante/a-z/informations-generales-sur-le-coronavirus/\",\"SK\":\"\",\"YT\":\"https://yukon.ca/fr/6-steps-staying-safe-covid-19\"},\"YourData\":{\"Body1\":\"votre identité\",\"Body2\":\"vos contacts\",\"Body3\":\"toute précision sur vous ou sur les personnes à proximité\",\"Body4\":\"vos informations de santé\",\"Body5\":\"\",\"Intro\":\"Alerte COVID n’a aucun moyen de connaître :\",\"Title\":\"Complètement anonyme\"}},\"pt-BR\":{\"Home\":{\"NoExposureDetected\":\"Nenhuma exposição detectada\",\"NoExposureDetectedDetailed\":\"Com base em seus IDs aleatórios, você não esteve perto de alguém nos últimos 14 dias que testou positivo para COVID-19.\",\"ExposureDetected\":\"Você possivelmente foi exposto ao COVID-19\",\"ExposureDetectedDetailed\":\"Com base nos seus IDs aleatórios, você esteve perto de alguém nos últimos 14 dias que testou positivo para COVID-19.\",\"SeeGuidance\":\"Leia a orientação do Canadá sobre COVID-19\",\"GuidanceUrl\":\"https://www.canada.ca/en/public-health/services/publications/diseases-conditions/covid-19-how-to-isolate-at-home.html\",\"SignalDataShared\":\"Obrigado por ajudar a diminuir a propagação\",\"SignalDataSharedDetailed\":{\"One\":\"Você receberá uma notificação solicitando que compartilhe seus IDs aleatórios todos os dias pelo próximo {number} dia.\",\"Other\":\"Você receberá uma notificação solicitando que compartilhe seus IDs aleatórios todos os dias pelos próximos {number} dias.\"},\"SignalDataSharedCTA\":\"Acompanhe seus sintomas\",\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DailyShare\":\"Compartilhe seus novos IDs aleatórios\",\"DailyShareDetailed\":\"Compartilhar seus dados ajuda outras pessoas a saber se elas foram expostas ao COVID-19.\",\"ShareRandomIDsCTA\":\"Compartilhe seus IDs aleatórios\",\"BluetoothDisabled\":\"Bluetooth desligado\",\"EnableBluetoothCTA\":\"O Bluetooth é usado para coletar e compartilhar IDs aleatórios necessários para o COVID Alert funcionar.\",\"TurnOnBluetooth\":\"Ligar o Bluetooth\",\"ExposureNotificationsDisabled\":\"As notificações de exposição estão desligadas\",\"ExposureNotificationsDisabledDetailed\":\"São necessárias notificações de exposição para que o COVID Alert funcione.\",\"EnableExposureNotificationsCTA\":\"Ligar notificações de exposição\",\"NoConnectivity\":\"Sem conexão à internet\",\"NoConnectivityDetailed\":\"O COVID Alert precisa de acesso à Internet para continuar verificando à potenciais exposições.\",\"AppName\":\"COVID Alert\",\"ExternalLinkHint\":\"Opens in a new window\",\"LastCheckedMinutes\":{\"One\":\"Última verificação há {number} minuto\",\"Other\":\"Última verificação há {number} minutos\"},\"LastCheckedHours\":{\"One\":\"Última verificação há {number} hora\",\"Other\":\"Última verificação há {number} horas\"},\"LastCheckedDays\":{\"One\":\"Última verificação há {number} dia\",\"Other\":\"Última verificação há {number} dias\"}},\"OverlayClosed\":{\"SystemStatus\":\"COVID Alert está \",\"SystemStatusOn\":\"LIGADO\",\"SystemStatusOff\":\"DESLIGADO\",\"NotificationStatus\":\"As notificações estão \",\"NotificationStatusOff\":\"DESLIGADAS\",\"TapPrompt\":\"Toque para mais informações\"},\"OverlayOpen\":{\"BluetoothCardStatus\":\"Bluetooth está \",\"BluetoothCardStatusOff\":\"DESLIGADO\",\"BluetoothCardBody\":\"O Bluetooth é usado para coletar e compartilhar IDs aleatórios necessários para o COVID Alert funcionar.\",\"BluetoothCardAction\":\"Ligar o Bluetooth\",\"EnterCodeCardTitle\":\"Código do COVID Alert\",\"EnterCodeCardBody\":\"Um profissional de saúde pediu para você inserir um código COVID Alert?\",\"EnterCodeCardAction\":\"Insira o código\",\"NotificationCardStatus\":\"As notificações estão \",\"NotificationCardStatusOff\":\"DESLIGADAS\",\"NotificationCardBody\":\"Você não receberá notificações se tiver sido exposto ao COVID-19. A notificação de potencial exposição é importante para ajudar a diminuir a propagação do COVID-19.\",\"NotificationCardAction\":\"Ligar notificações\",\"ExposureNotificationCardStatus\":\"As notificações de exposição estão \",\"ExposureNotificationCardStatusOff\":\"DESLIGADAS\",\"ExposureNotificationCardBody\":\"São necessárias notificações de exposição para que o COVID Alert funcione.\",\"ExposureNotificationCardAction\":\"Ligar notificações de exposição\"},\"Info\":{\"CheckSymptoms\":\"Verifique os sintomas\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/en\",\"TellAFriend\":\"Compartilhe o aplicativo\",\"LearnMore\":\"Como o COVID Alert funciona\",\"ChangeLanguage\":\"Mudar idioma\",\"Privacy\":\"Leia a política de privacidade\"},\"LanguageSelect\":{\"Title\":\"Idioma\",\"En\":\"Inglês (Canadá)\",\"Fr\":\"Francês (Canadá)\",\"PtBR\":\"Português (Brasil)\",\"EnShort\":\"Inglês\",\"FrShort\":\"Francês\",\"PtBRShort\":\"Português (Brasil)\",\"Close\":\"Voltar\"},\"Tutorial\":{\"step-1Title\":\"Como o COVID Alert funciona\",\"step-1\":\"Seu telefone usa o Bluetooth para coletar e compartilhar IDs aleatórios com telefones próximos a você com o COVID Alert instalado. Esses IDs são armazenados com segurança em cada telefone.\",\"step-2Title\":\"Se você foi potencialmente exposto ao COVID-19\",\"step-2\":\"Você receberá uma notificação se esteve próximo de alguém nos últimos 14 dias que testou positivo para COVID-19.\",\"step-3Title\":\"Se você testar positivo para COVID-19\",\"step-3\":\"Notifique anonimamente as pessoas que estiveram perto de você nos últimos 14 dias de que elas possivelmente foram expostas.\",\"ActionBack\":\"Voltar\",\"ActionNext\":\"Próximo\",\"ActionEnd\":\"Concluir\",\"Close\":\"Fechar\"},\"Sharing\":{\"Title\":\"Compartilhe o aplicativo\",\"SubTitle\":\"Todas as pessoas que usam o COVID Alert ajudam a diminuir a propagação do COVID-19. Compartilhando este aplicativo não compartilhará dados pessoais.\",\"Platform-messages\":\"Compartilhar em Mensagens\",\"Platform-instagram\":\"Compartilhar no Instagram\",\"More\":\"Mais aplicativos\",\"Message\":\"Junte-se a mim e ajude a diminuir a propagação do COVID-19. Baixe o aplicativo COVID Shield: https://covidshield.app\",\"InstagramImageUrl\":\"https://user-images.githubusercontent.com/5274722/84658989-b9ba6b00-aee4-11ea-84d4-d840527467b4.png\",\"Close\":\"Fechar\"},\"Onboarding\":{\"ActionNext\":\"Próximo\",\"ActionEnd\":\"Concluir\",\"ActionBack\":\"Voltar\",\"Start\":{\"Title\":\"Join the effort to slow the spread\",\"Body1\":\"COVID Alert notifies you if someone you’ve been near tells the app they tested positive.\",\"Body2\":\"That way, you can isolate right away and stop the spread.\",\"TutorialAction\":\"Learn how StopCOVID works\"},\"Anonymous\":{\"Title\":\"Totally Anonymous\",\"Body1\":\"COVID Alert does not use GPS or location services.\",\"Body2\":\"It has no way of knowing:\",\"Bullet1\":\"your location\",\"Bullet2\":\"your name or address\",\"Bullet3\":\"your phone's contacts\",\"Bullet4\":\"your health information\"},\"HowItWorks\":{\"Title\":\"How it works\",\"Body1\":\"The app uses Bluetooth to exchange random codes with nearby phones.\",\"Body2\":\"Every day, it checks a list of random codes from people who tell the app they tested positive.\",\"Body3\":\"If you’ve been near one of those codes in the past 14 days, you’ll get a notification.\",\"HowItWorksCTA\":\"Learn more about how it works\"},\"Permissions\":{\"Title\":\"The app will now ask your permission\",\"Body1\":\"Allow the app to start logging random codes or “random IDs” when you’re near other phones. The app will access the date, duration and signal strength related to the random codes, but they never leave your phone.\",\"Body2\":\"You’ll also need to enable push notifications.\",\"PrivacyButtonCTA\":\"Learn more about privacy\"}},\"Privacy\":{\"Title\":\"Política de privacidade\",\"Close\":\"Fechar\"},\"ThankYou\":{\"Title\":\"Obrigado por ajudar\",\"Body\":\"As notificações de exposição estão ativadas. Você receberá uma notificação se o COVID Alert detectar que você foi exposto ao COVID-19.\",\"Dismiss\":\"Dispensar\"},\"DataUpload\":{\"Cancel\":\"Cancelar\",\"FormIntro\":\"Digite seu código COVID Alert de 8 dígitos.\",\"Action\":\"Enviar código\",\"InputLabel\":\"StopCOVID code\",\"InfoSmall\":\"Seus IDs aleatórios não serão compartilhados, a menos que você permita na próxima etapa.\",\"ConsentTitle\":\"Compartilhe seus IDs aleatórios\",\"ConsentBody\":\"Você recebeu acesso de um profissional de saúde para compartilhar os IDs aleatórios armazenados no seu telefone.\",\"ConsentBody2\":\"Se você compartilhar seus IDs aleatórios, outras pessoas com a COVID Alert que estiveram perto de você nos últimos 14 dias receberão uma notificação de que elas possivelmente foram expostas. Esta notificação não incluirá nenhuma informação sobre você.\",\"ConsentBody3\":\"Você será notificado uma vez por dia quando novos IDs aleatórios estiverem disponíveis para serem compartilhados. Você deve conceder permissão cada vez para que os IDs aleatórios sejam compartilhados. Compartilhar é voluntário.\",\"PrivacyPolicyLink\":\"Ver política de privacidade\",\"ConsentAction\":\"Concordo\",\"ShareToast\":\"IDs aleatórios compartilhados com sucesso\",\"ErrorTitle\":\"Código não reconhecido\",\"ErrorBody\":\"Não foi possível reconhecer seu código COVID Alert. Por favor, tente novamente.\",\"ErrorAction\":\"OK\"},\"Notification\":{\"ExposedMessageTitle\":\"Você possivelmente foi exposto ao COVID-19\",\"ExposedMessageBody\":\"Com base nos seus IDs aleatórios, você esteve perto de alguém nos últimos 14 dias que testou positivo para COVID-19.\",\"OffMessageTitle\":\"O COVID Alert está desligado\",\"OffMessageBody\":\"Ligue o COVID Alert para ser notificado se você tiver sido potencialmente exposto ao COVID-19.\",\"DailyUploadNotificationTitle\":\"Compartilhe seus novos IDs aleatórios\",\"DailyUploadNotificationBody\":\"Compartilhar seus dados ajuda outras pessoas a saber se elas foram expostas ao COVID-19.\"},\"Partners\":{\"Label\":\"Desenvolvido em parceria com\"},\"BottomSheet\":{\"Collapse\":\"Ocultar\",\"OnStatus\":\"StopCOVID is on, Tap for more information\",\"OffStatus\":\"StopCOVID is off, Tap for more information\"},\"RegionPicker\":{\"Title\":\"Where do you live?\",\"Body\":\"Choose the province or territory whose public health guidance you need to follow.\\n\\nThis information is only stored on your phone and will never be shared with anyone else.\",\"Skip\":\"Skip\",\"GetStarted\":\"Get started\",\"AB\":\"Alberta\",\"BC\":\"British Columbia\",\"MB\":\"Manitoba\",\"NB\":\"New Brunswick\",\"NL\":\"Newfoundland and Labrador\",\"NT\":\"Northwest Territories\",\"NS\":\"Nova Scotia\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Prince Edward Island\",\"QC\":\"Quebec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\"},\"OnboardingPermissions\":{\"Title\":\"Seus dados são privados\",\"Body\":\"O COVID Alert adota uma abordagem de privacidade em primeiro lugar na notificação de exposição.\",\"Body2\":\"Seus IDs aleatórios são armazenados exclusivamente em seu telefone, a menos que você os compartilhe.\",\"Body3\":\"O uso do COVID Alert é voluntário. Você pode excluir seus IDs aleatórios a qualquer momento.\"},\"OnboardingStart\":{\"Title\":\"Mantenha você e os outros seguros\",\"Body1\":\"Seja notificado se você foi potencialmente exposto ao COVID-19.\",\"Body2\":\"Se você testar positivo para COVID-19, poderá compartilhar seus dados anonimamente para que outras pessoas possam ser notificadas sobre uma possível exposição.\",\"TutorialAction\":\"Saiba como o COVID Alert funciona\"}}}")
+export default JSON.parse("{\"en\":{\"Home\":{\"NoExposureDetected\":\"No exposure detected\",\"NoExposureDetectedDetailed\":\"You have not been near anyone who reported a COVID diagnosis through this app.\",\"ExposureDetected\":\"You’ve been exposed\",\"ExposureDetectedDetailed\":\"You’ve been near someone who reported a COVID-19 diagnosis through the app.\\n\\nThis does not mean you’ve been infected. But you should probably take some action.\",\"SeeGuidance\":\"What now? See guidance\",\"GuidanceUrl\":\"https://www.canada.ca/en/public-health/services/publications/diseases-conditions/covid-19-how-to-isolate-at-home.html\",\"SignalDataShared\":\"Thank you for helping to slow the spread\",\"SignalDataSharedDetailed\":{\"One\":\"For the next {number} day, you’ll get a daily message asking for your  information.\\n\\nTake care for yourself and your family during this time. \",\"Other\":\"For the next {number} days, you’ll get a daily message asking for your  information.\\n\\nTake care for yourself and your family during this time.\"},\"SignalDataSharedCTA\":\"Track your symptoms\",\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DailyShare\":\"Share your new information\",\"DailyShareDetailed\":\"Sharing your information helps others to know if they've been exposed to COVID-19. Then they can isolate and slow the spread to others. No details about who you are will be shared.\",\"ShareRandomIDsCTA\":\"Notify people you've been near\",\"BluetoothDisabled\":\"Bluetooth is off\",\"EnableBluetoothCTA\":\"You need to turn on Bluetooth in your phone’s settings so StopCOVID can work.\",\"TurnOnBluetooth\":\"Turn on Bluetooth\",\"ExposureNotificationsDisabled\":\"Exposure notifications are off\",\"ExposureNotificationsDisabledDetailed\":\"Exposure notifications are needed for StopCOVID to work.\",\"EnableExposureNotificationsCTA\":\"Turn on exposure notifications\",\"NoConnectivity\":\"No internet connection\",\"NoConnectivityDetailed\":\"StopCOVID needs internet access to continue checking for potential exposure.\",\"AppName\":\"StopCOVID\",\"ExternalLinkHint\":\"Opens in a new window\",\"LastCheckedMinutes\":{\"One\":\"Last checked {number} minute ago\",\"Other\":\"Last checked {number} minutes ago\"},\"LastCheckedHours\":{\"One\":\"Last checked {number} hour ago\",\"Other\":\"Last checked {number} hours ago\"},\"LastCheckedDays\":{\"One\":\"Last checked {number} day ago\",\"Other\":\"Last checked {number} days ago\"}},\"OverlayClosed\":{\"SystemStatus\":\"StopCOVID is \",\"SystemStatusOn\":\"ON\",\"SystemStatusOff\":\"OFF\",\"NotificationStatus\":\"Push notifications are \",\"NotificationStatusOff\":\"OFF\",\"TapPrompt\":\"Tap for more information\"},\"OverlayOpen\":{\"BluetoothCardStatus\":\"Bluetooth is \",\"BluetoothCardStatusOff\":\"OFF\",\"BluetoothCardBody\":\"You need to turn on Bluetooth in your phone’s settings so StopCOVID can work.\",\"BluetoothCardAction\":\"Turn on Bluetooth\",\"EnterCodeCardTitle\":\"Diagnosed with COVID-19?\",\"EnterCodeCardBody\":\"You were given an 8-digit number when you were diagnosed. This number tells StopCOVID to notify people you were near.\",\"EnterCodeCardAction\":\"Enter code\",\"NotificationCardStatus\":\"Push notifications are \",\"NotificationCardStatusOff\":\"OFF\",\"NotificationCardBody\":\"You will not receive push notifications if you've been near anyone who has reported a COVID-19 diagnosis. Finding out potential exposures right away is essential to slowing the spread of COVID-19.\",\"NotificationCardAction\":\"Turn on push notifications\",\"ExposureNotificationCardStatus\":\"Exposure notifications are \",\"ExposureNotificationCardStatusOff\":\"OFF\",\"ExposureNotificationCardBody\":\"Exposure notifications are needed for StopCOVID to work.\",\"ExposureNotificationCardAction\":\"Turn on exposure notifications\"},\"Info\":{\"CheckSymptoms\":\"Check symptoms\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/en\",\"TellAFriend\":\"Share app\",\"LearnMore\":\"How StopCOVID works\",\"ChangeLanguage\":\"Change language\",\"Privacy\":\"Read privacy policy\"},\"LanguageSelect\":{\"Title\":\"Language\",\"En\":\"English (Canada)\",\"Fr\":\"Français (Canada)\",\"PtBR\":\"Portuguese (Brazil)\",\"EnShort\":\"English\",\"FrShort\":\"Français\",\"PtBRShort\":\"Portuguese (Brazil)\",\"Close\":\"Back\"},\"Tutorial\":{\"step-1Title\":\"How StopCOVID works\",\"step-1\":\"Your phone uses Bluetooth to collect and share random IDs with phones near you with StopCOVID installed. These IDs are stored securely on each phone.\",\"step-2Title\":\"If you have potentially been exposed to COVID-19\",\"step-2\":\"You will receive a notification if you have been near someone in the past 14 days who has tested positive for COVID-19.\",\"step-3Title\":\"If you test positive for COVID-19\",\"step-3\":\"Anonymously notify others who have been near you in the past 14 days that they have possibly been exposed.\",\"ActionBack\":\"Back\",\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"Close\":\"Close\"},\"Sharing\":{\"Title\":\"Tell your friends about StopCOVID\",\"SubTitle\":\"Every person using StopCOVID helps to slow the spread of COVID-19. Sharing this app will not share any of your personal information.\",\"Platform-messages\":\"Share to Messages\",\"Platform-instagram\":\"Share to Instagram\",\"More\":\"More apps\",\"Message\":\"Join me in helping to slow the spread of COVID-19. Download the StopCOVID app: https://covidshield.app\",\"InstagramImageUrl\":\"https://user-images.githubusercontent.com/5274722/84658989-b9ba6b00-aee4-11ea-84d4-d840527467b4.png\",\"Close\":\"Close\"},\"Onboarding\":{\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"ActionBack\":\"Back\",\"Start\":{\"Title\":\"Join the effort to slow the spread\",\"Body1\":\"COVID Alert notifies you if someone you’ve been near tells the app they tested positive.\",\"Body2\":\"That way, you can isolate right away and stop the spread.\"},\"WhatItsNot\":{\"Title\":\"What COVID Alert does not do\",\"Body1\":\"The app will not tell you in the moment if you’re currently near someone who’s been diagnosed.\",\"Body2\":\"It will not tell you about outbreaks in your city or neighbourhood.\"},\"Anonymous\":{\"Title\":\"Totally Anonymous\",\"Body1\":\"COVID Alert does not use GPS or location services.\",\"Body2\":\"It has no way of knowing:\",\"Bullet1\":\"your location\",\"Bullet2\":\"your name or address\",\"Bullet3\":\"your phone's contacts\",\"Bullet4\":\"your health information\"},\"HowItWorks\":{\"Title\":\"How it works\",\"Body1\":\"The app uses Bluetooth to exchange random codes with nearby phones.\",\"Body2\":\"Every day, it checks a list of random codes from people who tell the app they tested positive.\",\"Body3\":\"If you’ve been near one of those codes in the past 14 days, you’ll get a notification.\",\"HowItWorksCTA\":\"Learn more about how it works\"},\"Permissions\":{\"Title\":\"The app will now ask your permission\",\"Body1\":\"Allow the app to start logging random codes or “random IDs” when you’re near other phones. The app will access the date, duration and signal strength related to the random codes, but they never leave your phone.\",\"Body2\":\"You’ll also need to enable push notifications.\",\"PrivacyButtonCTA\":\"Learn more about privacy\"}},\"Privacy\":{\"Title\":\"Privacy policy\",\"Close\":\"Close\"},\"ThankYou\":{\"Title\":\"Thank you for helping\",\"Body\":\"Exposure notifications are on. You will receive a message if StopCOVID detects that you've been near someone who has reported a COVID-19 diagnosis.\",\"Dismiss\":\"Dismiss\"},\"DataUpload\":{\"Cancel\":\"Cancel\",\"FormIntro\":\"Please enter your 8 digit StopCOVID code.\",\"Action\":\"Submit code\",\"InputLabel\":\"StopCOVID code\",\"InfoSmall\":\"Your random IDs will not be shared unless you give permission in the next step.\",\"ConsentTitle\":\"Share your random IDs\",\"ConsentBody\":\"You have been granted access by a health care professional to share the random IDs stored on your phone.\",\"ConsentBody2\":\"If you share your random IDs, others with StopCOVID who have been near you in the past 14 days will receive a notification that they have possibly been exposed. This notification will not include any information about you.\",\"ConsentBody3\":\"You will be notified once per day when new random IDs are available to be shared. You must grant permission each time for the random IDs to be shared. Sharing is voluntary.\",\"PrivacyPolicyLink\":\"View privacy policy\",\"ConsentAction\":\"I agree\",\"ShareToast\":\"Random IDs shared successfully\",\"ErrorTitle\":\"Code not recognized\",\"ErrorBody\":\"Your StopCOVID code could not be recognized. Please try again.\",\"ErrorAction\":\"OK\"},\"Notification\":{\"ExposedMessageTitle\":\"You’ve been exposed\",\"ExposedMessageBody\":\"You’ve been near someone who reported a COVID-19 diagnosis through the app.\\n\\nThis does not mean you’ve been infected. But you should probably take some action.\",\"OffMessageTitle\":\"StopCOVID is off\",\"OffMessageBody\":\"Turn on StopCOVID to get notified if you've been near someone who has reported a COVID-19 diagnosis.\",\"DailyUploadNotificationTitle\":\"Share your new informatin about people you were near\",\"DailyUploadNotificationBody\":\"Sharing your information helps people know if they've been exposed to COVID-19. Then they can isolate and slow the spread to others. No details about who you are will be shared.\"},\"Partners\":{\"Label\":\"Developed in partnership with\"},\"BottomSheet\":{\"Collapse\":\"Close\",\"OnStatus\":\"StopCOVID is on, Tap for more information\",\"OffStatus\":\"StopCOVID is off, Tap for more information\"},\"RegionPicker\":{\"Title\":\"Where do you live?\",\"Body\":\"Choose the province or territory whose public health guidance you need to follow.\\n\\nThis information is only stored on your phone and will never be shared with anyone else.\",\"Skip\":\"Skip\",\"GetStarted\":\"Get started\",\"AB\":\"Alberta\",\"BC\":\"British Columbia\",\"MB\":\"Manitoba\",\"NB\":\"New Brunswick\",\"NL\":\"Newfoundland and Labrador\",\"NT\":\"Northwest Territories\",\"NS\":\"Nova Scotia\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Prince Edward Island\",\"QC\":\"Quebec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\"}},\"fr\":{\"Home\":{\"NoExposureDetected\":{\"NoRegion\":{\"Body\":\"Vous n’avez été près d’aucune personne ayant signalé un diagnostic de COVID-19 avec cette application.\\n\\nToutefois, dans certaines provinces et certains territoires, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\",\"DiagnosedBtnText1\":\"Vous avez la COVID-19?\",\"DiagnosedBtnText2\":\"Prochaines étapes\",\"Title\":\"Aucune exposition détectée\"},\"RegionCovered\":{\"Body\":\"Vous n’avez été près d’aucune personne ayant signalé un diagnostic de COVID-19 avec cette application.\",\"DiagnosedBtnText1\":\"Vous avez la COVID-19?\",\"DiagnosedBtnText2\":\"Prochaines étapes\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"Title\":\"Aucune exposition détectée\"},\"RegionNotCovered\":{\"Body\":\"Les personnes de votre province ou territoire ne peuvent pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nIl est tout de même utile d’activer Alerte COVID. Quand les personnes pourront signaler un diagnostic, vous serez ainsi avisé si jamais vous avez été près d’elles.\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"Title\":\"Pas encore de signalement dans votre région\"}},\"NoExposureDetectedDetailed\":\"You have not been near anyone who reported a COVID diagnosis through this app.\",\"ExposureDetected\":{\"Detailed1\":\"Vous avez été près d’une personne qui a signalé un diagnostic de COVID-19 avec cette application.\",\"Detailed2\":\"Cela ne veut pas dire que vous êtes infecté. Mais vous devriez probablement prendre quelques mesures.\",\"DiagnosedBtnText1\":\"Vous avez la COVID-19?\",\"DiagnosedBtnText2\":\"Prochaines étapes\",\"Title\":\"Vous avez été exposé\"},\"ExposureDetectedDetailed\":\"You’ve been near someone who reported a COVID-19 diagnosis through the app.\\n\\nThis does not mean you’ve been infected. But you should probably take some action.\",\"SeeGuidance\":\"Que faire maintenant? Lisez les recommandations.\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"SignalDataShared\":\"Thank you for helping to slow the spread\",\"SignalDataSharedDetailed\":{\"One\":\"For the next {number} day, you’ll get a daily message asking for your  information.\\n\\nTake care for yourself and your family during this time. \",\"Other\":\"For the next {number} days, you’ll get a daily message asking for your  information.\\n\\nTake care for yourself and your family during this time.\"},\"SignalDataSharedCTA\":\"Track your symptoms\",\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DailyShare\":\"Partagez vos nouveaux identifiants aléatoires\",\"DailyShareDetailed\":\"Partager vos données aide les autres à savoir s'ils ont possiblement été exposés à la COVID-19.\",\"ShareRandomIDsCTA\":\"Partagez vos identifiants aléatoires\",\"BluetoothDisabled\":\"Bluetooth désactivé\",\"EnableBluetoothCTA\":\"Vous devez activer Bluetooth dans les réglages de votre téléphone pour que Alerte COVID fonctionne.\",\"TurnOnBluetooth\":\"Activez Bluetooth\",\"ExposureNotificationsDisabled\":\"Les notifications d'exposition à la COVID-19 sont désactivées\",\"ExposureNotificationsDisabledDetailed\":\"Les notifications d'exposition à la COVID-19 sont nécessaires au fonctionnement de Alerte COVID.\",\"EnableExposureNotificationsCTA\":\"Activez les notifications d'exposition à la COVID-19\",\"NoConnectivity\":\"Alerte COVID est hors ligne\",\"NoConnectivityDetailed\":\"Alerte COVID vérifie s’il y a des expositions seulement quand votre téléphone est connecté à Internet.\",\"AppName\":\"Alerte COVID\",\"ExternalLinkHint\":\"Ouvre dans une nouvelle fenêtre\",\"LastCheckedMinutes\":{\"One\":\"Dernière vérification il y a {number} minute\",\"Other\":\"Dernière vérification il y a {number} minutes\"},\"LastCheckedHours\":{\"One\":\"Dernière vérification il y a {number} heure\",\"Other\":\"Dernière vérification il y a {number} heures\"},\"LastCheckedDays\":{\"One\":\"Dernière vérification il y a {number} jour\",\"Other\":\"Dernière vérification il y a {number} jours\"},\"CTA\":\"Prochaines étapes\",\"ChooseRegionCTA\":\"Sélectionnez une province ou un territoire\",\"DiagnosedView\":{\"Body1\":{\"One\":\"Tomorrow, you’ll get your last notification asking permission to upload any new random codes your phone sent out.\",\"Other\":\"Au cours des {number} prochains jours, vous recevrez une notification quotidienne vous demandant l’autorisation de téléverser tout code aléatoire émis par votre téléphone.\"},\"Body2\":\"Vous n’êtes pas tenu d’accepter, même si vous avez accepté le jour d’avant. Vous avez le choix.\",\"Body3\":\"Prenez soin de vous et de votre famille pendant ce temps.\",\"Title\":\"Merci d’aider à lutter contre la COVID\"},\"How\":\"Diagnosed with COVID-19?\",\"HowUrl\":\"https://www.canada.ca/\"},\"OverlayClosed\":{\"SystemStatus\":\"Alerte COVID est \",\"SystemStatusOn\":\"ACTIVÉ\",\"SystemStatusOff\":\"DÉSACTIVÉ\",\"NotificationStatus\":\"Notifications \",\"NotificationStatusOff\":\"DÉSACTIVÉES\",\"TapPrompt\":\"Appuyez pour voir le menu\"},\"OverlayOpen\":{\"BluetoothCardStatus\":\"Bluetooth \",\"BluetoothCardStatusOff\":\"DÉSACTIVÉ\",\"BluetoothCardBody\":\"Vous devez l’activer dans les réglages de votre téléphone pour que Alerte COVID fonctionne.\",\"BluetoothCardAction\":\"Activer Bluetooth\",\"EnterCodeCardTitle\":\"Code Alerte COVID\",\"EnterCodeCardBody\":\"Un professionel de la santé vous a-t-il demandé d'entrer un code Alerte COVID?\",\"EnterCodeCardAction\":\"Entrez le code\",\"NotificationCardStatus\":\"Activer les notifications \",\"NotificationCardStatusOff\":\"DÉSACTIVÉES\",\"NotificationCardBody\":\"Vous ne recevrez pas de notifications poussées en cas d’exposition potentielle à la COVID-19. Le fait d’être rapidement notifié d’une exposition potentielle est important pour ralentir la propagation de COVID-19.\",\"NotificationCardAction\":\"Activer les notifications\",\"ExposureNotificationCardStatus\":\"Les notifications d'exposition à la COVID-19 sont \",\"ExposureNotificationCardStatusOff\":\"DÉSACTIVÉES\",\"ExposureNotificationCardBody\":\"Les notifications d'exposition à la COVID-19 sont nécessaires au fonctionnement de Alerte COVID.\",\"ExposureNotificationCardAction\":\"Activez les notifications d'exposition\",\"EnterCodeCardBodyDiagnosed\":\"Merci de faire votre part pour freiner la propagation du virus.\",\"EnterCodeCardTitleDiagnosed\":\"Vous contribuez à freiner la COVID\"},\"Info\":{\"CheckSymptoms\":\"Vérifier vos symptômes\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/fr\",\"TellAFriend\":\"Partager l’application\",\"LearnMore\":\"Fonctionnement de Alerte COVID\",\"ChangeLanguage\":\"Modifier la langue\",\"Privacy\":\"Lire la politique de confidentialité\",\"ChangeRegion\":\"Modifier la province\"},\"LanguageSelect\":{\"Title\":\"Langue\",\"En\":\"English (Canada)\",\"Fr\":\"Français (Canada)\",\"PtBR\":\"Portugais (Brésil)\",\"EnShort\":\"English\",\"FrShort\":\"Français\",\"PtBRShort\":\"Portugais (Brésil)\",\"Close\":\"Retour\"},\"Tutorial\":{\"step-1Title\":\"Fonctionnement de Alerte COVID\",\"step-1\":\"Après son installation, Alerte COVID s’exécute en arrière-plan sans interrompre vos activités.\\n\\nAlors que vous vaquez à vos activités, votre téléphone émet un code aléatoire par Bluetooth toutes les 5 minutes. L’application change fréquemment les codes. Ainsi, ils ne peuvent pas servir à vous identifier.\",\"step-2Title\":\"\",\"step-2\":\"Chaque fois qu’une personne ayant installé Alerte COVID se trouve près de vous, son téléphone émet des codes aléatoires que votre téléphone collecte et stocke. De même, l’autre téléphone collecte et stocke les codes aléatoires émis par votre téléphone.\\n\\nLes téléphones émettent des codes aléatoires toutes les 5 minutes. Ces codes sont stockés sur chacun des téléphones de façon sécurisée.\",\"step-3Title\":\"\",\"step-3\":\"L’application estime à quelle distance se trouvent les autres téléphones ayant Alerte COVID. Pour ce faire, elle utilise la puissance de leur signal Bluetooth.\\n\\nSi vous êtes à moins d’une certaine distance pendant plus d’un certain temps, l’application enregistre une exposition.\",\"ActionBack\":\"Précédent\",\"ActionNext\":\"Suivant\",\"ActionEnd\":\"Terminé\",\"Close\":\"Fermer\",\"step-4\":\"Si une personne ayant l’application a un diagnostic de COVID-19, elle peut décider de téléverser tous les codes émis par son téléphone. Les codes vont vers un serveur central.\\n\\nLe serveur ne reçoit que les codes. Il ne reçoit aucune information sur la personne.\",\"step-4Title\":\"\",\"step-5\":\"Chaque jour, votre téléphone vérifie le serveur quand il a une connexion Internet. Il obtient une liste de codes aléatoires émis par les téléphones des personnes qui ont signalé un diagnostic avec l’application.\\n\\nL’application supprime les codes collectés après 14 jours.\",\"step-5Title\":\"\",\"step-6\":\"L’application compare la liste de codes avec les codes aléatoires reçus d’autres téléphones.\\n\\nSi elle trouve des codes identiques, elle vérifie la puissance du signal Bluetooth et la durée pendant laquelle vous avez été à proximité. Si les codes sont associés à une exposition, l’application vous dira que vous avez peut-être été exposé.\\n\\nVous pouvez alors vous isoler et lire les recommandations de santé publique.\",\"step-6Title\":\"\"},\"Sharing\":{\"Title\":\"Partager l’application\",\"SubTitle\":\"Chaque personne utilisant Alerte COVID aide à ralentir la transmission de la COVID-19. Partager cette application ne partage aucune donnée privée.\",\"Platform-messages\":\"Partager avec Messages\",\"Platform-instagram\":\"Partager avec Instagram\",\"More\":\"Plus d'applications\",\"Message\":\"Joignez-vous à moi pour aider à ralentir la transmission de la COVID-19. Téléchargez l'application Alerte COVID : https://covidshield.app\",\"InstagramImageUrl\":\"https://user-images.githubusercontent.com/5274722/84658989-b9ba6b00-aee4-11ea-84d4-d840527467b4.png\",\"Close\":\"Fermer\"},\"Onboarding\":{\"ActionNext\":\"Suivant\",\"ActionEnd\":\"Terminé\",\"ActionBack\":\"Précédent\",\"Start\":{\"Title\":\"Join the effort to slow the spread\",\"Body1\":\"COVID Alert notifies you if someone you’ve been near tells the app they tested positive.\",\"Body2\":\"That way, you can isolate right away and stop the spread.\"},\"WhatItsNot\":{\"Title\":\"What COVID Alert does not do\",\"Body1\":\"The app will not tell you in the moment if you’re currently near someone who’s been diagnosed.\",\"Body2\":\"It will not tell you about outbreaks in your city or neighbourhood.\"},\"Anonymous\":{\"Title\":\"Totally Anonymous\",\"Body1\":\"COVID Alert does not use GPS or location services.\",\"Body2\":\"It has no way of knowing:\",\"Bullet1\":\"your location\",\"Bullet2\":\"your name or address\",\"Bullet3\":\"your phone's contacts\",\"Bullet4\":\"your health information\"},\"HowItWorks\":{\"Title\":\"How it works\",\"Body1\":\"The app uses Bluetooth to exchange random codes with nearby phones.\",\"Body2\":\"Every day, it checks a list of random codes from people who tell the app they tested positive.\",\"Body3\":\"If you’ve been near one of those codes in the past 14 days, you’ll get a notification.\",\"HowItWorksCTA\":\"Learn more about how it works\"},\"Permissions\":{\"Title\":\"The app will now ask your permission\",\"Body1\":\"Allow the app to start logging random codes or “random IDs” when you’re near other phones. The app will access the date, duration and signal strength related to the random codes, but they never leave your phone.\",\"Body2\":\"You’ll also need to enable push notifications.\",\"PrivacyButtonCTA\":\"Learn more about privacy\"}},\"Privacy\":{\"Title\":\"Politique de confidentialité\",\"Close\":\"Fermer\"},\"ThankYou\":{\"Title\":\"Merci d'aider\",\"Body\":\"Les notifications d’exposition sont activées. Si Alerte COVID détecte que vous avez possiblement été exposé(e) à la COVID-19, vous recevrez une notification.\",\"Dismiss\":\"OK\"},\"DataUpload\":{\"Cancel\":\"Annuler\",\"FormIntro\":\"Entrez votre numéro à 8 chiffres.\",\"Action\":\"Envoyer\",\"InputLabel\":\"code Alerte COVID\",\"InfoSmall\":\"Vos identifiants aléatoires ne seront pas partagés, à moins que vous donniez votre permission à l'étape suivante.\",\"ConsentTitle\":\"Permettre à Alerte COVID d’aviser les autres personnes\",\"ConsentBody\":\"You have been granted access by a health care professional to share the random IDs stored on your phone.\",\"ConsentBody2\":\"Personne ne recevra d’information sur vous ou sur le moment où vous avez été en proximité.\",\"ConsentBody3\":\"Il s’agit de votre choix. Vous pouvez aider à ralentir la propagation de COVID-19 si vous acceptez.\",\"PrivacyPolicyLink\":\"Comment Alerte COVID protège votre vie privée\",\"ConsentAction\":\"Accepter\",\"ShareToast\":\"Identifiants aléatoires partagés avec succès\",\"ErrorTitle\":\"Code non reconnu\",\"ErrorBody\":\"Votre code Alerte COVID n’a pas pu être reconnu. Veuillez essayer à nouveau.\",\"ErrorAction\":\"OK\",\"ConsentBody1\":\"Alerte COVID vous demandera l’autorisation de téléverser vos codes aléatoires des 14 derniers jours.\",\"ConsentSubtitle\":\"C’est votre choix\",\"HowItWorksLink\":\"Comment Alerte COVID fonctionne\",\"Step1\":{\"Body\":\"Vous devriez avoir reçu un numéro à 8 chiffres quand vous avez eu votre diagnostic. Vous devez entrer ce numéro si vous voulez que Alerte COVID avise les personnes qu’elles ont été exposées.\\n\\nEntrer le code est optionnel. Vous avez le choix. Mais vous pouvez aider à ralentir la propagation de COVID-19 si vous le faites.\",\"CTA\":\"Entrer le numéro\",\"NoCode\":\"Vous n’avez pas de numéro à 8 chiffres?\",\"Title\":\"Entrez votre numéro à 8 chiffres\"},\"NoCode\":{\"NoRegion\":{\"Title\":\"Vous n’avez pas choisi de province ou de territoire\",\"Body\":\"Dans certaines provinces et certains territoires, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nPour savoir si vous pouvez obtenir un numéro à 8 chiffres, vous devez sélectionner votre région.\",\"ChooseRegionCTA\":\"Sélectionner une province ou un territoire\"},\"RegionCovered\":{\"Title\":\"Vous n’avez pas de numéro à 8 chiffres?  \",\"Body\":\"Contactez la personne ou le bureau qui vous a donné vos résultats afin d’obtenir un numéro.\"},\"RegionNotCovered\":{\"Title\":\"Pas encore de signalement dans votre région\",\"Body\":\"Les personnes de votre province ou territoire ne peuvent pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nVous ne pourrez pas aviser les personnes dont vous avez été à proximité qu’elles ont été exposées. Mais vous pouvez tout de même prendre des mesures pour :\",\"Body2\":\"assurer la sécurité de votre communauté;\",\"Body3\":\"protéger les personnes et les animaux vivant avec vous;\",\"Body4\":\"prendre soin de vous.\"}}},\"Notification\":{\"ExposedMessageTitle\":\"Vous avez possiblement été exposé(e)\",\"ExposedMessageBody\":\"Selon vos identifiants aléatoires, au cours des 14 derniers jours, vous avez été à proximité d'une personne ayant reçu un résultat de test positif à la COVID-19.\",\"OffMessageTitle\":\"Alerte COVID est désactivé\",\"OffMessageBody\":\"Activez Alerte COVID pour recevoir une notification si vous avez potentiellement été exposé(e) à la COVID-19.\",\"DailyUploadNotificationTitle\":\"Partagez vos nouveaux identifiants aléatoires\",\"DailyUploadNotificationBody\":\"Partager vos données aide les autres à savoir s'ils ont possiblement été exposés à la COVID-19.\"},\"Partners\":{\"Label\":\"Conçu en partenariat avec\"},\"BottomSheet\":{\"Collapse\":\"Fermer\",\"OnStatus\":\"Alerte COVID est ACTIVÉ, Appuyez pour voir le menu\",\"OffStatus\":\"Alerte COVID est DÉSACTIVÉ, Appuyez pour voir le menu\"},\"RegionPicker\":{\"Title\":\"Où habitez-vous?\",\"Body\":\"Sélectionnez la province ou le territoire dont vous devez suivre les recommandations de santé publique.\\n\\nCette information est stockée sur votre téléphone uniquement et ne sera jamais transmise à d’autres personnes.\",\"Skip\":\"Ignorer\",\"GetStarted\":\"Commencer\",\"AB\":\"Alberta\",\"BC\":\"Colombie-Britannique\",\"MB\":\"Manitoba\",\"NB\":\"Nouveau-Brunswick\",\"NL\":\"Terre-Neuve-et-Labrador\",\"NT\":\"Territoires du Nord-Ouest\",\"NS\":\"Nouvelle-Écosse\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Île-du-Prince-Édouard\",\"QC\":\"Québec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\",\"Close\":\"Fermer\",\"Optional\":\"* Facultatif\",\"SettingsTitle\":\"Province ou territoire\"},\"OnboardingLocation\":{\"Body1\":\"votre position\",\"Body2\":\"votre identité\",\"Body3\":\"les endroits où vous êtes allé\",\"Body4\":\"les personnes avec qui vous avez été\",\"Body5\":\"Vous pouvez le supprimer à tout moment\",\"Intro\":\"Alerte COVID n’utilise pas le GPS ou les services de localisation. Il n’a aucun moyen de connaître :\",\"Title\":\"Aucun suivi de position\"},\"OnboardingPermissions\":{\"Body1\":\"Rien n’est fait sans votre autorisation.\",\"Body2\":\"Même quand vous donnez votre autorisation, Alerte COVID ne partage que les codes aléatoires des téléphones dont vous avez été à proximité. Il ne partage aucune information sur vous.\",\"HowItWorksCTA\":\"Lire sur le fonctionnement de Alerte COVID\",\"PrivacyButtonCTA\":\"Lire la politique de confidentialité\",\"Title\":\"Aucune action sans votre autorisation\"},\"OnboardingStart\":{\"Body1\":\"Il vous avise si au cours des 14 derniers jours, vous avez été à proximité d’une personne ayant signalé un diagnostic de COVID-19 avec l’application.\",\"Body2\":\"Il ne vous dira pas en temps réel si vous êtes à proximité d’une personne ayant eu un diagnostic.\",\"Body3\":\"If you test positive for COVID-19, you can anonymously notify others who have been near you, in the past 14 days.\",\"Body4\":\"Vous n’êtes pas tenu de l’utiliser\",\"Body5\":\"Vous pouvez le supprimer à tout moment\",\"Title\":\"Ce que fait Alerte COVID\",\"Title2\":\"Vous avez le choix\",\"TutorialAction\":\"Découvrez comment Alerte COVID fonctionne\"},\"RegionalGuidanceURL\":{\"AB\":\"\",\"BC\":\"\",\"CA\":\"https://www.canada.ca/fr/sante-publique/services/maladies/2019-nouveau-coronavirus/symptomes.html\",\"MB\":\"https://www.gov.mb.ca/covid19/about/index.fr.html\",\"NB\":\"https://www2.gnb.ca/content/gnb/fr/corporate/promo/covid-19.html\",\"NL\":\"\",\"NS\":\"https://novascotia.ca/coronavirus/when-to-seek-help/fr/\",\"NT\":\"https://www.gov.nt.ca/covid-19/fr/services/%C3%A0-propos-de-la-covid-19\",\"NU\":\"https://www.gov.nu.ca/fr/sante/information/covid-19-nouveau-coronavirus-0\",\"ON\":\"https://covid-19.ontario.ca/fr/centres-devaluation-de-la-covid-19\",\"PE\":\"https://www.princeedwardisland.ca/fr/information/sante-et-mieux-etre/covid-19-foire-aux-questions\",\"QC\":\"https://www.quebec.ca/sante/problemes-de-sante/a-z/informations-generales-sur-le-coronavirus/\",\"SK\":\"\",\"YT\":\"https://yukon.ca/fr/6-steps-staying-safe-covid-19\"},\"YourData\":{\"Body1\":\"votre identité\",\"Body2\":\"vos contacts\",\"Body3\":\"toute précision sur vous ou sur les personnes à proximité\",\"Body4\":\"vos informations de santé\",\"Body5\":\"\",\"Intro\":\"Alerte COVID n’a aucun moyen de connaître :\",\"Title\":\"Complètement anonyme\"}},\"pt-BR\":{\"Home\":{\"NoExposureDetected\":\"Nenhuma exposição detectada\",\"NoExposureDetectedDetailed\":\"Com base em seus IDs aleatórios, você não esteve perto de alguém nos últimos 14 dias que testou positivo para COVID-19.\",\"ExposureDetected\":\"Você possivelmente foi exposto ao COVID-19\",\"ExposureDetectedDetailed\":\"Com base nos seus IDs aleatórios, você esteve perto de alguém nos últimos 14 dias que testou positivo para COVID-19.\",\"SeeGuidance\":\"Leia a orientação do Canadá sobre COVID-19\",\"GuidanceUrl\":\"https://www.canada.ca/en/public-health/services/publications/diseases-conditions/covid-19-how-to-isolate-at-home.html\",\"SignalDataShared\":\"Obrigado por ajudar a diminuir a propagação\",\"SignalDataSharedDetailed\":{\"One\":\"Você receberá uma notificação solicitando que compartilhe seus IDs aleatórios todos os dias pelo próximo {number} dia.\",\"Other\":\"Você receberá uma notificação solicitando que compartilhe seus IDs aleatórios todos os dias pelos próximos {number} dias.\"},\"SignalDataSharedCTA\":\"Acompanhe seus sintomas\",\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DailyShare\":\"Compartilhe seus novos IDs aleatórios\",\"DailyShareDetailed\":\"Compartilhar seus dados ajuda outras pessoas a saber se elas foram expostas ao COVID-19.\",\"ShareRandomIDsCTA\":\"Compartilhe seus IDs aleatórios\",\"BluetoothDisabled\":\"Bluetooth desligado\",\"EnableBluetoothCTA\":\"O Bluetooth é usado para coletar e compartilhar IDs aleatórios necessários para o COVID Alert funcionar.\",\"TurnOnBluetooth\":\"Ligar o Bluetooth\",\"ExposureNotificationsDisabled\":\"As notificações de exposição estão desligadas\",\"ExposureNotificationsDisabledDetailed\":\"São necessárias notificações de exposição para que o COVID Alert funcione.\",\"EnableExposureNotificationsCTA\":\"Ligar notificações de exposição\",\"NoConnectivity\":\"Sem conexão à internet\",\"NoConnectivityDetailed\":\"O COVID Alert precisa de acesso à Internet para continuar verificando à potenciais exposições.\",\"AppName\":\"COVID Alert\",\"ExternalLinkHint\":\"Opens in a new window\",\"LastCheckedMinutes\":{\"One\":\"Última verificação há {number} minuto\",\"Other\":\"Última verificação há {number} minutos\"},\"LastCheckedHours\":{\"One\":\"Última verificação há {number} hora\",\"Other\":\"Última verificação há {number} horas\"},\"LastCheckedDays\":{\"One\":\"Última verificação há {number} dia\",\"Other\":\"Última verificação há {number} dias\"}},\"OverlayClosed\":{\"SystemStatus\":\"COVID Alert está \",\"SystemStatusOn\":\"LIGADO\",\"SystemStatusOff\":\"DESLIGADO\",\"NotificationStatus\":\"As notificações estão \",\"NotificationStatusOff\":\"DESLIGADAS\",\"TapPrompt\":\"Toque para mais informações\"},\"OverlayOpen\":{\"BluetoothCardStatus\":\"Bluetooth está \",\"BluetoothCardStatusOff\":\"DESLIGADO\",\"BluetoothCardBody\":\"O Bluetooth é usado para coletar e compartilhar IDs aleatórios necessários para o COVID Alert funcionar.\",\"BluetoothCardAction\":\"Ligar o Bluetooth\",\"EnterCodeCardTitle\":\"Código do COVID Alert\",\"EnterCodeCardBody\":\"Um profissional de saúde pediu para você inserir um código COVID Alert?\",\"EnterCodeCardAction\":\"Insira o código\",\"NotificationCardStatus\":\"As notificações estão \",\"NotificationCardStatusOff\":\"DESLIGADAS\",\"NotificationCardBody\":\"Você não receberá notificações se tiver sido exposto ao COVID-19. A notificação de potencial exposição é importante para ajudar a diminuir a propagação do COVID-19.\",\"NotificationCardAction\":\"Ligar notificações\",\"ExposureNotificationCardStatus\":\"As notificações de exposição estão \",\"ExposureNotificationCardStatusOff\":\"DESLIGADAS\",\"ExposureNotificationCardBody\":\"São necessárias notificações de exposição para que o COVID Alert funcione.\",\"ExposureNotificationCardAction\":\"Ligar notificações de exposição\"},\"Info\":{\"CheckSymptoms\":\"Verifique os sintomas\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/en\",\"TellAFriend\":\"Compartilhe o aplicativo\",\"LearnMore\":\"Como o COVID Alert funciona\",\"ChangeLanguage\":\"Mudar idioma\",\"Privacy\":\"Leia a política de privacidade\"},\"LanguageSelect\":{\"Title\":\"Idioma\",\"En\":\"Inglês (Canadá)\",\"Fr\":\"Francês (Canadá)\",\"PtBR\":\"Português (Brasil)\",\"EnShort\":\"Inglês\",\"FrShort\":\"Francês\",\"PtBRShort\":\"Português (Brasil)\",\"Close\":\"Voltar\"},\"Tutorial\":{\"step-1Title\":\"Como o COVID Alert funciona\",\"step-1\":\"Seu telefone usa o Bluetooth para coletar e compartilhar IDs aleatórios com telefones próximos a você com o COVID Alert instalado. Esses IDs são armazenados com segurança em cada telefone.\",\"step-2Title\":\"Se você foi potencialmente exposto ao COVID-19\",\"step-2\":\"Você receberá uma notificação se esteve próximo de alguém nos últimos 14 dias que testou positivo para COVID-19.\",\"step-3Title\":\"Se você testar positivo para COVID-19\",\"step-3\":\"Notifique anonimamente as pessoas que estiveram perto de você nos últimos 14 dias de que elas possivelmente foram expostas.\",\"ActionBack\":\"Voltar\",\"ActionNext\":\"Próximo\",\"ActionEnd\":\"Concluir\",\"Close\":\"Fechar\"},\"Sharing\":{\"Title\":\"Compartilhe o aplicativo\",\"SubTitle\":\"Todas as pessoas que usam o COVID Alert ajudam a diminuir a propagação do COVID-19. Compartilhando este aplicativo não compartilhará dados pessoais.\",\"Platform-messages\":\"Compartilhar em Mensagens\",\"Platform-instagram\":\"Compartilhar no Instagram\",\"More\":\"Mais aplicativos\",\"Message\":\"Junte-se a mim e ajude a diminuir a propagação do COVID-19. Baixe o aplicativo COVID Shield: https://covidshield.app\",\"InstagramImageUrl\":\"https://user-images.githubusercontent.com/5274722/84658989-b9ba6b00-aee4-11ea-84d4-d840527467b4.png\",\"Close\":\"Fechar\"},\"Onboarding\":{\"ActionNext\":\"Próximo\",\"ActionEnd\":\"Concluir\",\"ActionBack\":\"Voltar\",\"Start\":{\"Title\":\"Join the effort to slow the spread\",\"Body1\":\"COVID Alert notifies you if someone you’ve been near tells the app they tested positive.\",\"Body2\":\"That way, you can isolate right away and stop the spread.\"},\"WhatItsNot\":{\"Title\":\"What COVID Alert does not do\",\"Body1\":\"The app will not tell you in the moment if you’re currently near someone who’s been diagnosed.\",\"Body2\":\"It will not tell you about outbreaks in your city or neighbourhood.\"},\"Anonymous\":{\"Title\":\"Totally Anonymous\",\"Body1\":\"COVID Alert does not use GPS or location services.\",\"Body2\":\"It has no way of knowing:\",\"Bullet1\":\"your location\",\"Bullet2\":\"your name or address\",\"Bullet3\":\"your phone's contacts\",\"Bullet4\":\"your health information\"},\"HowItWorks\":{\"Title\":\"How it works\",\"Body1\":\"The app uses Bluetooth to exchange random codes with nearby phones.\",\"Body2\":\"Every day, it checks a list of random codes from people who tell the app they tested positive.\",\"Body3\":\"If you’ve been near one of those codes in the past 14 days, you’ll get a notification.\",\"HowItWorksCTA\":\"Learn more about how it works\"},\"Permissions\":{\"Title\":\"The app will now ask your permission\",\"Body1\":\"Allow the app to start logging random codes or “random IDs” when you’re near other phones. The app will access the date, duration and signal strength related to the random codes, but they never leave your phone.\",\"Body2\":\"You’ll also need to enable push notifications.\",\"PrivacyButtonCTA\":\"Learn more about privacy\"}},\"Privacy\":{\"Title\":\"Política de privacidade\",\"Close\":\"Fechar\"},\"ThankYou\":{\"Title\":\"Obrigado por ajudar\",\"Body\":\"As notificações de exposição estão ativadas. Você receberá uma notificação se o COVID Alert detectar que você foi exposto ao COVID-19.\",\"Dismiss\":\"Dispensar\"},\"DataUpload\":{\"Cancel\":\"Cancelar\",\"FormIntro\":\"Digite seu código COVID Alert de 8 dígitos.\",\"Action\":\"Enviar código\",\"InputLabel\":\"StopCOVID code\",\"InfoSmall\":\"Seus IDs aleatórios não serão compartilhados, a menos que você permita na próxima etapa.\",\"ConsentTitle\":\"Compartilhe seus IDs aleatórios\",\"ConsentBody\":\"Você recebeu acesso de um profissional de saúde para compartilhar os IDs aleatórios armazenados no seu telefone.\",\"ConsentBody2\":\"Se você compartilhar seus IDs aleatórios, outras pessoas com a COVID Alert que estiveram perto de você nos últimos 14 dias receberão uma notificação de que elas possivelmente foram expostas. Esta notificação não incluirá nenhuma informação sobre você.\",\"ConsentBody3\":\"Você será notificado uma vez por dia quando novos IDs aleatórios estiverem disponíveis para serem compartilhados. Você deve conceder permissão cada vez para que os IDs aleatórios sejam compartilhados. Compartilhar é voluntário.\",\"PrivacyPolicyLink\":\"Ver política de privacidade\",\"ConsentAction\":\"Concordo\",\"ShareToast\":\"IDs aleatórios compartilhados com sucesso\",\"ErrorTitle\":\"Código não reconhecido\",\"ErrorBody\":\"Não foi possível reconhecer seu código COVID Alert. Por favor, tente novamente.\",\"ErrorAction\":\"OK\"},\"Notification\":{\"ExposedMessageTitle\":\"Você possivelmente foi exposto ao COVID-19\",\"ExposedMessageBody\":\"Com base nos seus IDs aleatórios, você esteve perto de alguém nos últimos 14 dias que testou positivo para COVID-19.\",\"OffMessageTitle\":\"O COVID Alert está desligado\",\"OffMessageBody\":\"Ligue o COVID Alert para ser notificado se você tiver sido potencialmente exposto ao COVID-19.\",\"DailyUploadNotificationTitle\":\"Compartilhe seus novos IDs aleatórios\",\"DailyUploadNotificationBody\":\"Compartilhar seus dados ajuda outras pessoas a saber se elas foram expostas ao COVID-19.\"},\"Partners\":{\"Label\":\"Desenvolvido em parceria com\"},\"BottomSheet\":{\"Collapse\":\"Ocultar\",\"OnStatus\":\"StopCOVID is on, Tap for more information\",\"OffStatus\":\"StopCOVID is off, Tap for more information\"},\"RegionPicker\":{\"Title\":\"Where do you live?\",\"Body\":\"Choose the province or territory whose public health guidance you need to follow.\\n\\nThis information is only stored on your phone and will never be shared with anyone else.\",\"Skip\":\"Skip\",\"GetStarted\":\"Get started\",\"AB\":\"Alberta\",\"BC\":\"British Columbia\",\"MB\":\"Manitoba\",\"NB\":\"New Brunswick\",\"NL\":\"Newfoundland and Labrador\",\"NT\":\"Northwest Territories\",\"NS\":\"Nova Scotia\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Prince Edward Island\",\"QC\":\"Quebec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\"},\"OnboardingPermissions\":{\"Title\":\"Seus dados são privados\",\"Body\":\"O COVID Alert adota uma abordagem de privacidade em primeiro lugar na notificação de exposição.\",\"Body2\":\"Seus IDs aleatórios são armazenados exclusivamente em seu telefone, a menos que você os compartilhe.\",\"Body3\":\"O uso do COVID Alert é voluntário. Você pode excluir seus IDs aleatórios a qualquer momento.\"},\"OnboardingStart\":{\"Title\":\"Mantenha você e os outros seguros\",\"Body1\":\"Seja notificado se você foi potencialmente exposto ao COVID-19.\",\"Body2\":\"Se você testar positivo para COVID-19, poderá compartilhar seus dados anonimamente para que outras pessoas possam ser notificadas sobre uma possível exposição.\",\"TutorialAction\":\"Saiba como o COVID Alert funciona\"}}}")
